Create Utils class
